requirements.txt
# Core dependencies (for modular version)
sqlalchemy>=1.4.0
click>=8.0.0
tabulate>=0.8.0
python-dotenv>=0.19.0

# Development dependencies
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0